refactor: Move remaining time calculation to featureTimerStore, remove optimistic start
Timer no longer starts counting down optimistically — state and lastActionDate
are set only after the server responds to start(). The remaining/interval logic
moves from apiTimerStore to featureTimerStore, which also fixes the display
flicker on pause (remaining is preserved from the last tick rather than reset
to the stale durationLeft).
